Build mysql as shared extension for PHP 5.5 and 5.6.
Updated Dockerfiles for PHP 5.5 and 5.6 to build mysql as a shared
extension and disable it by default.

The mysql extension is deprecated since PHP 5.5 and using it is
discouraged. Users requiring the mysql extension can still enable it
by adding;

extension=mysql.so

to their php.ini, which should be added to the README because this
breaks backwards compatibility with previous versions of this image.

After building, the mysql extension can be found here;

/usr/local/lib/php/extensions/no-debug-non-zts-20131226/mysql.so

closes #28